requirements.txt
Django==4.1.6
django-money==3.0.0
stripe==5.1.1
python-dotenv==0.21.1
